refactors entities from script_thread into script_runtime
