+++ components/script/dom/filereader.rs
@@ -22,8 +22,9 @@ use encoding::label::encoding_from_whatwg_label;
 use encoding::types::{DecoderTrap, EncodingRef};
 use hyper::mime::{Attr, Mime};
 use rustc_serialize::base64::{CharacterSet, Config, Newline, ToBase64};
-use script_thread::ScriptThreadEventCategory::FileRead;
-use script_thread::{CommonScriptMsg, Runnable, ScriptChan};
+use script_runtime::ScriptThreadEventCategory::FileRead;
+use script_runtime::{ScriptChan, CommonScriptMsg};
+use script_thread::Runnable;
 use std::cell::Cell;
 use string_cache::Atom;
 use util::str::DOMString;
