+++ components/script/dom/htmldetailselement.rs
@@ -14,7 +14,8 @@ use dom::eventtarget::EventTarget;
 use dom::htmlelement::HTMLElement;
 use dom::node::{Node, window_from_node};
 use dom::virtualmethods::VirtualMethods;
-use script_thread::{MainThreadScriptChan, Runnable, ScriptChan};
+use script_runtime::ScriptChan;
+use script_thread::{MainThreadScriptChan, Runnable};
 use std::cell::Cell;
 use string_cache::Atom;
 use task_source::dom_manipulation::DOMManipulationTask;
