+++ components/script/dom/htmlimageelement.rs
@@ -25,8 +25,9 @@ use ipc_channel::ipc;
 use ipc_channel::router::ROUTER;
 use net_traits::image::base::{Image, ImageMetadata};
 use net_traits::image_cache_thread::{ImageResponder, ImageResponse};
-use script_thread::ScriptThreadEventCategory::UpdateReplacedElement;
-use script_thread::{CommonScriptMsg, Runnable, ScriptChan};
+use script_runtime::ScriptThreadEventCategory::UpdateReplacedElement;
+use script_runtime::{CommonScriptMsg, ScriptChan};
+use script_thread::Runnable;
 use std::sync::Arc;
 use string_cache::Atom;
 use url::Url;
