+++ components/script/dom/htmlinputelement.rs
@@ -32,8 +32,9 @@ use dom::validation::Validatable;
 use dom::virtualmethods::VirtualMethods;
 use msg::constellation_msg::ConstellationChan;
 use range::Range;
-use script_thread::ScriptThreadEventCategory::InputEvent;
-use script_thread::{CommonScriptMsg, Runnable};
+use script_runtime::CommonScriptMsg;
+use script_runtime::ScriptThreadEventCategory::InputEvent;
+use script_thread::Runnable;
 use script_traits::ScriptMsg as ConstellationMsg;
 use std::borrow::ToOwned;
 use std::cell::Cell;
