+++ components/script/dom/htmlscriptelement.rs
@@ -34,7 +34,8 @@ use js::jsapi::RootedValue;
 use js::jsval::UndefinedValue;
 use net_traits::{AsyncResponseListener, AsyncResponseTarget, Metadata};
 use network_listener::{NetworkListener, PreInvoke};
-use script_thread::{MainThreadScriptChan, ScriptChan};
+use script_runtime::ScriptChan;
+use script_thread::MainThreadScriptChan;
 use std::ascii::AsciiExt;
 use std::cell::Cell;
 use std::mem;
