+++ components/script/dom/servohtmlparser.rs
@@ -29,7 +29,8 @@ use msg::constellation_msg::{PipelineId, SubpageId};
 use net_traits::{AsyncResponseListener, Metadata};
 use network_listener::PreInvoke;
 use parse::Parser;
-use script_thread::{ScriptChan, ScriptThread};
+use script_runtime::ScriptChan;
+use script_thread::ScriptThread;
 use std::cell::Cell;
 use std::cell::UnsafeCell;
 use std::default::Default;
