+++ components/script/dom/storage.rs
@@ -16,7 +16,8 @@ use dom::urlhelper::UrlHelper;
 use ipc_channel::ipc;
 use net_traits::storage_thread::{StorageThread, StorageThreadMsg, StorageType};
 use page::IterablePage;
-use script_thread::{MainThreadRunnable, MainThreadScriptChan, ScriptChan, ScriptThread};
+use script_runtime::ScriptChan;
+use script_thread::{MainThreadRunnable, MainThreadScriptChan, ScriptThread};
 use task_source::dom_manipulation::DOMManipulationTask;
 use url::Url;
 use util::str::DOMString;
