+++ components/script/dom/websocket.rs
@@ -35,8 +35,9 @@ use net_traits::MessageData;
 use net_traits::hosts::replace_hosts;
 use net_traits::unwrap_websocket_protocol;
 use net_traits::{WebSocketCommunicate, WebSocketConnectData, WebSocketDomAction, WebSocketNetworkEvent};
-use script_thread::ScriptThreadEventCategory::WebSocketEvent;
-use script_thread::{CommonScriptMsg, Runnable, ScriptChan};
+use script_runtime::ScriptThreadEventCategory::WebSocketEvent;
+use script_runtime::{CommonScriptMsg, ScriptChan};
+use script_thread::Runnable;
 use std::ascii::AsciiExt;
 use std::borrow::ToOwned;
 use std::cell::Cell;
