+++ components/script/network_listener.rs
@@ -3,8 +3,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use net_traits::{AsyncResponseListener, ResponseAction};
-use script_thread::ScriptThreadEventCategory::NetworkEvent;
-use script_thread::{CommonScriptMsg, Runnable, ScriptChan};
+use script_runtime::ScriptThreadEventCategory::NetworkEvent;
+use script_runtime::{CommonScriptMsg, ScriptChan};
+use script_thread::Runnable;
 use std::sync::{Arc, Mutex};
 
 /// An off-thread sink for async network event runnables. All such events are forwarded to
