+++ components/script/task_source/history_traversal.rs
@@ -2,7 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use script_thread::{CommonScriptMsg, MainThreadScriptMsg, ScriptChan};
+use script_runtime::{ScriptChan, CommonScriptMsg};
+use script_thread::MainThreadScriptMsg;
 use std::sync::mpsc::Sender;
 
 #[derive(JSTraceable)]
